Initial commit
Seems to be working. Did make some changes as some things seemed
a little odd. Ran into problems with running the tests, and after
far too long I realise that Powershell just doesn't like using
the -- to separate arguments, which needs to be quoted in order
to work correctly. Don't know what I was using earlier that did
not require that to be quoted.