Combine frontend and backend build
Added backend using static builds of the frontend, which I guess
would be the necessary way of doing it here as we're putting both
in the same repo. Not sure if I should change something anywhere,
as maybe some communication is now not necessary, but it does seem
to work, anyway.